backend/requirements.txt
fastapi==0.109.2
uvicorn==0.34.0
SQLAlchemy==2.0.40
psycopg2-binary==2.9.10
python-dotenv==1.0.1
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-jose==3.5.0
email-validator==2.1.1
python-multipart==0.0.20
openai==2.32.0
python-docx==1.2.0
pypdf==3.17.4
pdfplumber==0.11.9
langdetect==1.0.9
stripe==15.1.0
resend==2.29.0